v4

+++ kernel/syscall.c
@@ -103,7 +103,7 @@ extern uint64 sys_sleep(void);
 extern uint64 sys_unlink(void);
 extern uint64 sys_wait(void);
 extern uint64 sys_write(void);
-extern uint64 sys_uptime(void);
+extern uint  sys_uptime(void);
 extern uint64 sys_getprocs(void);
 extern uint64 sys_wait2(void);
 extern uint64 sys_getpriority(void);
